+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - New #125: Add `subscriptionPeriod` field to `TransactionPartnerUser` type.
 - New #125: Add `setUserEmojiStatus` method.
 - New #125: Add `PreparedInlineMessage` type.
+- New #125: Add `SavePreparedInlineMessage` method.
 
 ## 0.4.5 October 31, 2024